pca dir

+++ exp/ntu_hcc/move_profile.py
@@ -11,13 +11,18 @@ def copy_sra_files(csv_file, source_dir, dest_dir):
     with open(csv_file, 'r') as file:
         reader = csv.DictReader(file)
         for row in reader:
-            sra_ids.append(row['sra_id'])
+            sra_ids.append(row['sample'])
 
     # Copy files
     for sra_id in sra_ids:
         source_file = os.path.join(source_dir, sra_id, f"{sra_id}.tsv")
         dest_file = os.path.join(dest_dir, f"{sra_id}.tsv")
-        
+       
+        # Check if the file already exists in the destination
+        if os.path.exists(dest_file):
+            print(f"Skipping {sra_id}.tsv: File already exists in destination")
+            continue
+
         try:
             shutil.copy2(source_file, dest_file)
             print(f"Copied {sra_id}.tsv successfully")
@@ -32,7 +37,5 @@ def copy_sra_files(csv_file, source_dir, dest_dir):
 csv_file = 'all_family.csv'
 source_dir = '/mnt/d'
 dest_dir = '.'  # Current directory
-
 copy_sra_files(csv_file, source_dir, dest_dir)
-
 print("File copying process completed.")
